modified attack random attack block attack function move
