+++ client/src/battle/cardEffect/all/attack.ts
@@ -1,11 +1,11 @@
 import { CardEffectProps } from '../../../types/battle/cardEffect'
-import { playerAttack } from '../../../battle/player'
+import { attack } from '../../../common/cardEffect'
 import { setDamage } from '../../../battle/enemy'
 
 export const strike = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = playerAttack(player, enemy, card)
+    const damage = attack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
