+++ client/src/battle/cardEffect/attacker/attack.ts
@@ -1,12 +1,12 @@
 import { CardEffectProps } from '../../../types/battle/cardEffect'
 import { addBlock } from '../../../common/cardEffect'
-import { playerAttack, playerBlockAttack } from '../../../battle/player'
+import { attack, blockAttack } from '../../../common/cardEffect'
 import { setDamage } from '../../../battle/enemy'
 
 export const scorpion = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = playerAttack(player, enemy, card)
+    const damage = attack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
@@ -14,7 +14,7 @@ export const scorpion = (props: CardEffectProps): void => {
 export const moleclaw = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = playerAttack(player, enemy, card)
+    const damage = attack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
@@ -23,7 +23,7 @@ export const mantis = (props: CardEffectProps): void => {
   if (props.type === "allAttack") {
     const { player, enemies, card } = props
     enemies.forEach(enemy => {
-      const damage = playerAttack(player, enemy, card)
+      const damage = attack(player, enemy, card)
       setDamage(enemy, damage)
     })
   }
@@ -32,7 +32,7 @@ export const mantis = (props: CardEffectProps): void => {
 export const sword = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = playerAttack(player, enemy, card)
+    const damage = attack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
@@ -40,7 +40,7 @@ export const sword = (props: CardEffectProps): void => {
 export const raygust = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = playerAttack(player, enemy, card)
+    const damage = attack(player, enemy, card)
     setDamage(enemy, damage)
     addBlock(player, card.defense)
   }
@@ -49,8 +49,8 @@ export const raygust = (props: CardEffectProps): void => {
 export const thruster = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    playerBlockAttack(enemy, card)
-    const damage = playerAttack(player, enemy, card)
+    blockAttack(enemy, card)
+    const damage = attack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
