+++ client/src/battle/cardEffect/shooter/attack.ts
@@ -1,11 +1,11 @@
 import { CardEffectProps } from '../../../types/battle/cardEffect'
-import { playerAttack, randomPlayerAttack } from '../../../battle/player'
+import { attack, randomAttack } from '../../../common/cardEffect'
 import { setDamage } from '../../../battle/enemy'
 
 export const asteroid = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = randomPlayerAttack(player, enemy, card)
+    const damage = randomAttack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
@@ -13,7 +13,7 @@ export const asteroid = (props: CardEffectProps): void => {
 export const meteora = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = playerAttack(player, enemy, card)
+    const damage = attack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
@@ -21,7 +21,7 @@ export const meteora = (props: CardEffectProps): void => {
 export const viper = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = randomPlayerAttack(player, enemy, card)
+    const damage = randomAttack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
@@ -29,7 +29,7 @@ export const viper = (props: CardEffectProps): void => {
 export const hound = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = playerAttack(player, enemy, card)
+    const damage = attack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
@@ -38,7 +38,7 @@ export const grenade = (props: CardEffectProps): void => {
   if (props.type === "allAttack") {
     const { player, enemies, card } = props
     enemies.forEach(enemy => {
-      const damage = playerAttack(player, enemy, card)
+      const damage = attack(player, enemy, card)
       setDamage(enemy, damage)
     })
   }
