+++ client/src/battle/cardEffect/sniper/attack.ts
@@ -1,11 +1,11 @@
 import { CardEffectProps } from '../../../types/battle/cardEffect'
-import { playerAttack } from '../../../battle/player'
+import { attack } from '../../../common/cardEffect'
 import { setDamage } from '../../../battle/enemy'
 
 export const eagred = (props: CardEffectProps): void => {
   if (props.type === "oneAttack") {
     const { player, enemy, card } = props
-    const damage = playerAttack(player, enemy, card)
+    const damage = attack(player, enemy, card)
     setDamage(enemy, damage)
   }
 }
@@ -14,7 +14,7 @@ export const lightning = (props: CardEffectProps): void => {
   if (props.type === "allAttack") {
     const { player, enemies, card } = props
     enemies.forEach(enemy => {
-      const damage = playerAttack(player, enemy, card)
+      const damage = attack(player, enemy, card)
       setDamage(enemy, damage)
     })
   }
@@ -24,7 +24,7 @@ export const ibis = (props: CardEffectProps): void => {
   if (props.type === "allAttack") {
     const { player, enemies, card } = props
     enemies.forEach(enemy => {
-      const damage = playerAttack(player, enemy, card)
+      const damage = attack(player, enemy, card)
       setDamage(enemy, damage)
     })
   }
